added boxplot to eda
